Split data_utils up
